+++ src/vault/vault.h
@@ -13,13 +13,13 @@ class Vault{
         explicit Vault(std::string studentFilePath, std::string courseFilePath, std::string collegeFilePath);
 
         std::unordered_map<uint32_t,Student> LoadStudents();
-        std::unordered_map<uint16_t, College> LoadColleges();
-        std::unordered_map<uint16_t, Program> LoadCourses();
+        std::unordered_map<uint32_t, College> LoadColleges();
+        std::unordered_map<uint32_t, Program> LoadCourses();
 
 
         bool saveStudents(const std::unordered_map<uint32_t,Student>& );
-        bool savePrograms(const std::unordered_map<uint16_t, Program>&);
-        bool saveColleges(const std::unordered_map<uint16_t, College>&);
+        bool savePrograms(const std::unordered_map<uint32_t, Program>&);
+        bool saveColleges(const std::unordered_map<uint32_t, College>&);
     private:
         Gender ParseGender(std::string field);
         char SerializeGender(Gender gender);
